chore: init utils core
