+++ src/scripts/core.ts
@@ -1,8 +1,8 @@
-import { execa } from 'execa'
-import gradient from 'gradient-string'
-import fg from 'fast-glob'
 import { readFileSync, writeFileSync } from 'fs'
+import { execa } from 'execa'
+import fg from 'fast-glob'
 const { sync: glob } = fg
+import { logger, writeConsoleMsgs } from './utils'
 import {
   CheckFiles,
   ConstructVersionMapOptions,
@@ -11,40 +11,7 @@ import {
   PackageJSON,
   DepToUpdateItem,
   DepsToUpdate,
-  LoggerParams,
 } from '../types'
-
-/**
- * logger
- * @description logs to console messages
- * @param {LoggerParams.type} string
- * @param {LoggerParams.section} string
- * @param {LoggerParams.message} string
- * @param {LoggerParams.err} string
- * @returns {void}
- */
-export const logger = ({ type, section = '', message, err = '', isDebugging = false }: LoggerParams): void => {
-  const emoji = `🤼‍♀️`
-  const gap = ` => `
-  const debugMsg = isDebugging ? 'debugging:' : ''
-  const sectionMsg = section.length ? `${section}:` : ''
-  const firstLine = `codependence:${debugMsg}${sectionMsg}`
-  const secondLine = message ? `${emoji}${gap}${message}` : ''
-  if (type === 'error') {
-    console.error(gradient.passion(firstLine))
-    if (secondLine) console.error(secondLine)
-    if (err) console.error(err)
-  } else if (type === 'debug') {
-    console.debug(gradient.passion(firstLine))
-    if (secondLine) console.debug(secondLine)
-  } else if (type === 'info') {
-    console.info(gradient.teen(firstLine))
-    if (secondLine) console.info(secondLine)
-  } else {
-    console.log(gradient.teen(firstLine))
-    if (secondLine) console.log(secondLine)
-  }
-}
 
 /**
  * constructVersionMap
@@ -161,24 +128,6 @@ export const constructDepsToUpdateList = (
       exact: versionMap[name],
       expected: `${bumpCharacter}${versionMap[name]}`,
     }))
-}
-
-/**
- * writeConsoleMsgs
- * @param {packageName} string
- * @param {depList} array
- * @returns void
- */
-export const writeConsoleMsgs = (packageName: string, depList: Array<Record<string, string>>): void => {
-  if (!depList.length) return
-  Array.from(depList, ({ name: depName, expected, actual }) => {
-    logger({
-      type: 'log',
-      section: packageName,
-      message: `${depName} version is incorrect!`,
-    })
-    console.log(`🤼‍♀️ => Found ${actual} and should be ${expected}`)
-  })
 }
 
 /**
@@ -421,4 +370,5 @@ export const checkFiles = async ({
 
 export const script = checkFiles
 export const codependence = checkFiles
-export default codependence
+export const core = codependence
+export default core
